extensions/command-validation/pyproject.toml
[project]
authors = [
  {email = "engineering@canvasmedical.com", name = "Canvas Team"}
]
dependencies = [
  "canvas[test-utils]",
  "pytest-cov>=7.0.0",
]
description = "Some description of your project."
license = "MIT"
name = "command-validation"
readme = "command_validation/README.md"
requires-python = ">=3.11"
version = "0.1.0"

[tool.pytest.ini_options]
python_files = ["*_tests.py", "test_*.py", "tests.py"]
